@@ -2,10 +2,10 @@ apply plugin: 'com.android.library'
 group = 'com.github.ybq'
 version = '1.0'
 android {
-    compileSdkVersion 33
+    compileSdkVersion 34
     defaultConfig {
-        minSdkVersion 23
-        targetSdkVersion 33
+        minSdkVersion 26
+        targetSdkVersion 34
     }
     namespace 'com.github.ybq.android.spinkit'
 }
