@@ -4,10 +4,10 @@ group = 'com.ramotion.foldingcell'
 version = '1.1.0'
 
 android {
-    compileSdkVersion 33
+    compileSdkVersion 34
     defaultConfig {
-        minSdkVersion 23
-        targetSdkVersion 33
+        minSdkVersion 26
+        targetSdkVersion 34
     }
     buildTypes {
         release {
@@ -37,9 +37,8 @@ dependencies {
 }
 
 
-
-
-task jacocoReport(type: JacocoReport, dependsOn: 'testDebugUnitTest') {
+tasks.register('jacocoReport', JacocoReport) {
+    dependsOn 'testDebugUnitTest'
     group = 'Reporting'
     description = 'Generate Jacoco coverage reports after running tests.'
 
