@@ -16,20 +16,11 @@ public class OptAnimationLoader {
     public static Animation loadAnimation(Context context, int id)
             throws Resources.NotFoundException {
 
-        XmlResourceParser parser = null;
-        try {
-            parser = context.getResources().getAnimation(id);
+        try (XmlResourceParser parser = context.getResources().getAnimation(id)) {
             return createAnimationFromXml(context, parser);
-        } catch (XmlPullParserException ex) {
-            Resources.NotFoundException rnf = new Resources.NotFoundException("Can't load animation resource ID #0x" +
+        } catch (XmlPullParserException | IOException ex) {
+            throw new Resources.NotFoundException("Can't load animation resource ID #0x" +
                     Integer.toHexString(id), ex);
-            throw rnf;
-        } catch (IOException ex) {
-            Resources.NotFoundException rnf = new Resources.NotFoundException("Can't load animation resource ID #0x" +
-                    Integer.toHexString(id), ex);
-            throw rnf;
-        } finally {
-            if (parser != null) parser.close();
         }
     }
 
@@ -57,23 +48,30 @@ public class OptAnimationLoader {
 
             String  name = parser.getName();
 
-            if (name.equals("set")) {
-                anim = new AnimationSet(c, attrs);
-                createAnimationFromXml(c, parser, (AnimationSet)anim, attrs);
-            } else if (name.equals("alpha")) {
-                anim = new AlphaAnimation(c, attrs);
-            } else if (name.equals("scale")) {
-                anim = new ScaleAnimation(c, attrs);
-            }  else if (name.equals("rotate")) {
-                anim = new RotateAnimation(c, attrs);
-            }  else if (name.equals("translate")) {
-                anim = new TranslateAnimation(c, attrs);
-            } else {
-                try {
-                    anim = (Animation) Class.forName(name).getConstructor(Context.class, AttributeSet.class).newInstance(c, attrs);
-                } catch (Exception te) {
-                    throw new RuntimeException("Unknown animation name: " + parser.getName() + " error:" + te.getMessage());
-                }
+            switch (name) {
+                case "set":
+                    anim = new AnimationSet(c, attrs);
+                    createAnimationFromXml(c, parser, (AnimationSet) anim, attrs);
+                    break;
+                case "alpha":
+                    anim = new AlphaAnimation(c, attrs);
+                    break;
+                case "scale":
+                    anim = new ScaleAnimation(c, attrs);
+                    break;
+                case "rotate":
+                    anim = new RotateAnimation(c, attrs);
+                    break;
+                case "translate":
+                    anim = new TranslateAnimation(c, attrs);
+                    break;
+                default:
+                    try {
+                        anim = (Animation) Class.forName(name).getConstructor(Context.class, AttributeSet.class).newInstance(c, attrs);
+                    } catch (Exception te) {
+                        throw new RuntimeException("Unknown animation name: " + parser.getName() + " error:" + te.getMessage());
+                    }
+                    break;
             }
 
             if (parent != null) {
